Get rid of AIX linker option: -bnorwexec
High-level: this linker option is oxymoronic with JIT environment (where executable code are generated on the fly anyway).

At least for the time being before we figured out why default codeCache is allocated through malloc (instead of mmap or shmget, for which that linker option appeared to have no effect by my tests).

Even after we figured it out eventually, it is a legitimate question what value this option provides for JIT environment.

this also has effect on FFI thunks (which are not generated in JIT codecache).

Issue ibmruntimes/openj9-openjdk-jdk21#162